requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.7
django-celery-results==2.5.1
django-environ==0.11.2
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
kombu==5.4.2
prompt_toolkit==3.0.50
psycopg2==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
